used to_f instead .0 in #11

+++ ex12.rb
@@ -6,7 +6,7 @@ puts "A bigger number is #{bigger}"
 
 print 'Give me another numeber: '
 another = gets.chomp
-number = another.to_i
+number = another.to_f
 
-smaller = number / 100.0
+smaller = number / 100
 puts "A smaller number is #{smaller}"
